# packages/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SYNAPSE API Gateway",
    "version": "1.0.0-s2",
    "description": "Atlas Console BFF surface. Hand-curated S2 snapshot covering /auth, /api/v1/{decisions,orders,agents,rum,stream}. The CI gate `tests/contract/test_openapi_byte_equality.py` keeps this in lockstep with the live FastAPI app (B5)."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Local dev (via reverse-proxy nginx on port 80)"
    }
  ],
...
  "paths": {
...
    "/api/v1/decisions/": {
      "post": {
        "tags": ["decisions"],
        "operationId": "submitDecision",
        "summary": "Trigger a consensus decision (proxied to orchestrator)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DecisionRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Decision result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DecisionResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "503": { "$ref": "#/components/responses/UpstreamUnavailable" }
        },
        "security": [{ "sessionCookie": [], "csrfHeader": [] }]
      }
    },
    "/api/v1/decisions/recent": {
      "get": {
        "tags": ["decisions"],
        "operationId": "recentDecisions",
        "summary": "Recent decisions (paged via opaque cursor — S3)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 20 }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque keyset cursor returned in `next_cursor`.",
            "schema": { "type": ["string", "null"], "minLength": 1, "maxLength": 200 }
          },
          {
            "name": "tier",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/Tier" }
          },
          {
            "name": "escalated",
            "in": "query",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent rows",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecentDecisionsResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "503": { "$ref": "#/components/responses/UpstreamUnavailable" }
        },
        "security": [{ "sessionCookie": [] }]
      }
    },
    "/api/v1/decisions/{decision_id}": {
      "get": {
        "tags": ["decisions"],
        "operationId": "getDecision",
        "summary": "Full audit_consensus row for one decision (B2)",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Full audit_consensus row",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConsensusDecision" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/UpstreamUnavailable" }
        },
        "security": [{ "sessionCookie": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "synapse_session",
        "description": "Opaque session id minted by /auth/login. HttpOnly + Secure + SameSite=Strict."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF",
        "description": "Double-submit token returned at login. Required on POST/PUT/PATCH/DELETE."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPError" } } }
      },
      "Unauthorized": {
        "description": "Missing / expired session, or CSRF check failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPError" } } }
      },
      "NotFound": {
        "description": "No such resource",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPError" } } }
      },
      "UpstreamUnavailable": {
        "description": "An upstream service (orchestrator / Postgres / Kafka) is unreachable",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPError" } } }
      }
    },
    "schemas": {
      "HTTPError": {
        "type": "object",
        "required": ["detail"],
        "properties": { "detail": { "type": "string" } }
      },
...
      "City": {
        "type": "string",
        "enum": ["bengaluru", "mumbai"]
      },
      "Tier": {
        "type": "string",
        "enum": ["tier_1", "tier_2", "tier_3", "tier_4"]
      },
...
      "DecisionRequest": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "order_id": { "type": ["string", "null"], "format": "uuid" },
          "store_id": { "type": ["string", "null"] },
          "items": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "agents_involved": { "type": "array", "items": { "type": "string" } },
          "disruption_active": { "type": "boolean" },
          "requires_twin_simulation": { "type": "boolean" },
          "city": { "$ref": "#/components/schemas/City" }
        }
      },
      "DecisionResponse": {
        "type": "object",
        "required": ["decision_id", "tier", "confidence", "phase_reached", "escalated"],
        "properties": {
          "decision_id": { "type": "string", "format": "uuid" },
          "tier": { "$ref": "#/components/schemas/Tier" },
          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
          "phase_reached": { "type": "integer", "minimum": 1, "maximum": 5 },
          "audit_id": { "type": ["string", "null"], "format": "uuid" },
          "escalated": { "type": "boolean" }
        }
      },
      "RecentDecisionRow": {
        "type": "object",
        "required": ["audit_id", "decision_id", "tier", "created_at"],
        "properties": {
          "audit_id": { "type": "string", "format": "uuid" },
          "decision_id": { "type": "string", "format": "uuid" },
          "tier": { "$ref": "#/components/schemas/Tier" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "RecentDecisionsResponse": {
        "type": "object",
        "required": ["decisions", "count", "next_cursor"],
        "properties": {
          "decisions": { "type": "array", "items": { "$ref": "#/components/schemas/RecentDecisionRow" } },
          "count": { "type": "integer", "minimum": 0 },
          "next_cursor": {
            "type": ["string", "null"],
            "description": "Opaque cursor for the next page; null when no more rows."
          }
        }
      },
      "ConsensusDecision": {
        "type": "object",
        "required": [
          "id",
          "decision_id",
          "timestamp",
          "tier",
          "phase_reached",
          "proposals",
          "selected_action",
          "pareto_weights",
          "confidence",
          "debate_rounds",
          "escalated",
          "execution_confirmations",
          "context_messages",
          "audit_trace",
          "created_at"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "decision_id": { "type": "string", "format": "uuid" },
          "timestamp": { "type": "string", "format": "date-time" },
          "tier": { "$ref": "#/components/schemas/Tier" },
          "phase_reached": { "type": "integer", "minimum": 1, "maximum": 5 },
          "proposals": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "selected_action": { "type": "object", "additionalProperties": true },
          "pareto_weights": { "type": "object", "additionalProperties": true },
          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
          "debate_rounds": { "type": "integer", "minimum": 0 },
          "escalated": { "type": "boolean" },
          "human_override": { "type": ["object", "null"], "additionalProperties": true },
          "execution_confirmations": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "context_messages": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "audit_trace": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "pareto_front": { "type": ["array", "null"], "items": { "type": "object", "additionalProperties": true } },
          "outcome": { "type": ["object", "null"], "additionalProperties": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": []
}